Test data for complicated tone re-ordering
The difficulty comes with a consonant cluster,
which would form a conjunct, but not if there
is a nukta. If the matra is a left vowel, such as
U+0B47 ORIYA VOWEL SIGN E), the tone needs
to follow the vowel sign. If the tone is the overline,
the shaping engines have been updated and this
works. But if the tone is one of the two recently
added tones, the shaping engine does not do the
needed re-ordering.